+++ src/utils/index.ts
@@ -3,6 +3,8 @@ export { BaseParticle } from "./BaseParticle";
 export type { Body } from "./Body";
 export { BoxBody, PolygonBody } from "./Body";
 export { DynamicProperty } from "./DynamicProperty";
+export { EventTarget } from "./EventTarget";
+export { EventTargetObjectState } from "./EventTargetObjectState";
 export { MapSet } from "./MapSet";
 export { MergeProperty } from "./MergeProperty";
 export { ObjectState } from "./ObjectState";